requirements.txt
flask
google-genai
python-telegram-bot
python-dotenv
werkzeug
